Basic refactoring of class names and organization
