+++ Assets/Scripts/Cutting/FadeText.cs
